backend/scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.1
fake-useragent==1.4.0
